Debugging the thread pool
