+++ src/tasks/mod.rs
@@ -1,9 +1,5 @@
 mod pool;
-pub(crate) mod pool2;
 mod task_manager;
-mod worker_handle;
+mod worker;
 
-pub(crate) use self::{
-    pool::{schedule_task, RunCommand, ThreadPool},
-    task_manager::TaskManager,
-};
+pub(crate) use self::{pool::ThreadPool, task_manager::TaskManager};
